wechat-mcp/requirements.txt
pyautogui>=0.9.54
pygetwindow>=0.0.9
Pillow>=12.1.0
pyperclip>=1.11.0
